+++ frontend/src/pages/Sales.jsx
@@ -27,15 +27,19 @@ const Sales = () => {
   };
 
   const Delete = async (id) => {
-    try {
-      const response = await axios.delete(
-        `http://localhost:3000/api/sales/${id}`
-      );
-      console.log(response.data.message);
-      fetchSales();
-      toast.success(response.data.message);
-    } catch (error) {
-      console.error(error);
+    const confirmed = confirm("Do you want to delete this sale?");
+    if (confirmed) {
+      try {
+        const response = await axios.delete(
+          `http://localhost:3000/api/sales/${id}`
+        );
+
+        console.log(response.data.message);
+        fetchSales();
+        toast.success(response.data.message);
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
